+++ src/design/plone/contenttypes/tests/test_ct_bando.py
@@ -2,8 +2,10 @@
 from design.plone.contenttypes.testing import (
     DESIGN_PLONE_CONTENTTYPES_INTEGRATION_TESTING,
 )
-from redturtle.bandi.interfaces.settings import IBandoSettings
 from plone import api
+from plone.app.testing import setRoles
+from plone.app.testing import TEST_USER_ID
+from redturtle.bandi.interfaces.settings import IBandoSettings
 
 import unittest
 
@@ -14,6 +16,9 @@ class TestBando(unittest.TestCase):
     def setUp(self):
         """Custom shared utility setup for tests."""
         self.portal = self.layer["portal"]
+        self.request = self.layer["request"]
+        self.portal_url = self.portal.absolute_url()
+        setRoles(self.portal, TEST_USER_ID, ["Manager"])
 
     def test_bando_folder_deepening_addable_types(self):
         portal_types = api.portal.get_tool(name="portal_types")
@@ -32,3 +37,11 @@ class TestBando(unittest.TestCase):
             "default_ente", interface=IBandoSettings
         )
         self.assertEqual(default_ente, ())
+
+    def test_bando_substructure_created(self):
+
+        bando = api.content.create(container=self.portal, type="Bando", title="Bando")
+
+        self.assertIn("documenti", bando.keys())
+        self.assertIn("comunicazioni", bando.keys())
+        self.assertIn("esiti", bando.keys())